imp: 统一 redis/CachedHashMap 的超时时间为: 秒

+++ Web/src/main/java/org/voovan/http/server/SessionManager.java
@@ -53,7 +53,7 @@ public class SessionManager{
 
 		if(expirePutMethod == null) {
 			try {
-				expirePutMethod = TReflect.findMethod(httpSessions.getClass(), "put", String.class, String.class, int.class);
+				expirePutMethod = TReflect.findMethod(httpSessions.getClass(), "put", String.class, String.class, long.class);
 			} catch (ReflectiveOperationException e) {
 				expirePutMethod = null;
 			}
